update codedeploy config for Dockerfile

+++ scripts/prep-deploy.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
 
-source /etc/serviceapiprofile
+# Create a fresh application environment
+touch /etc/serviceapienv
+chmod 600 /etc/serviceapienv
+chown serviceapi:serviceapi /etc/serviceapienv
+bash /home/serviceapi/app/scripts/appenv.sh > /etc/serviceapienv
+if [[ $? -ne 0 ]]; then
+	exit 1
+fi
 
-if [[ -f /home/serviceapi/app/ansible/container.yml ]]; then
-	(cd /home/serviceapi/app/ansible && ansible-container stop)
+if [[ -f /home/serviceapi/app/Dockerfile ]]; then
+	docker rm $(docker stop $(docker ps -a -q --filter ancestor=serviceapi:latest))
 fi
 
 rm -rf /home/serviceapi/app
